Version bump of minor number.
Managed to squeeze a lot of the main updates into it (primarily text
colour changes).  Some changes from the main version could not be
included due to the 100,000 byte limit of userstyles.org, but it now
gives a slightly better indication of what the real theme does.  This
version retains the ICCF sponsor banner/link (which is pretty discreet
anyway), whereas the larger/main version hides it.

Unfortunately it will be very hard to include future updates and stay
within the size limit given the current file size.