A branch for an alternative approach to the solution.
Instead of modifying the PNG files for the chess board, leave them
entirely as they are.  The only thing we are attempting to change is
the coordinates around the edge edges of the board.  So the user can
switch off the coordinates in the ICCF interface and we can add the
relevant CSS and HTML to make those letters and numbers appear in the
size, font and colour of our choice.  This is a better solution
because it is more adaptable and preserves the ability to claim that
there is no interference with the board at all.

Even so, we may as well keep local copies of the board PNG files,
either for faster loading ot for reference.